src/requirements.txt
Flask>=3.1.0
numpy>=2.2.0
pandas>=2.2.3
python-dotenv>=1.0.1
pytz>=2023.3.post1
scipy>=1.14.1
tabulate>=0.9.0
twitchAPI>=4.3.1
